new release (including now FS-1091) (#49)
Create a new release in the `releases` folder.
Also add for the latest release the standalong chapters (with correct
section numbering and updated links).
From here, the full doc can be used to create a pdf version, and the
chapters can be used to create a web site.

Also disabling the linter for now, as it still produces line length
errors.