add spec web site and contributor's guide (#59)
With this PR, we produce a proper web site for the spec rather than just
a large markdown file.
The PR also fixes a number of markdown and link issues that became
visible thereby.
Readme and front matter have been adapted.
And a contributor's guide was added.
All of this should make us ready to attract more contributors.